requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
langgraph
chromadb
pypdf
python-dotenv
pydantic